api/requirements.txt
pinecone-client>=3.0.0
pinecone-plugin-assistant>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=22.0.0
flask-cors>=4.0.0